Moved RenderUtil + cleaned imports
