+++ src/main/java/mod/render360/gui/Render360Settings.java
@@ -1,12 +1,8 @@
 package mod.render360.gui;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.List;
 
-import core.render360.coretransform.RenderUtil;
+import mod.render360.RenderUtil;
 import mod.render360.render.RenderMethod;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiScreen;
